ciclos vida componentes
