+++ src/app/myfirst/myfirst.component.html
@@ -1,3 +1,5 @@
+<div style="border: 2px solid red">
 <p>
-myfirst works yeah MDF!!
+myfirst works yeah {{name}} MDF!!!
 </p>
+</div>
